# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste API",
        "description": "Teste LIVRO",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Lista todos os assuntos",
                "description": "Retorna uma coleção paginada de assuntos",
                "operationId": "c855f23389ed099cf68e0ec8b726d707",
                "responses": {
                    "200": {
                        "description": "Lista de assuntos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "cod_as": {
                                                "description": "Código do Assunto",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "descricao": {
                                                "description": "Descrição",
                                                "type": "string",
                                                "example": "Terror"
                                            },
                                            "created_at": {
                                                "description": "Data e hora de criação do registro",
                                                "type": "string",
                                                "example": "01/01/2025 12:00:00"
                                            },
                                            "updated_at": {
                                                "description": "Data e hora de atualização do registro",
                                                "type": "string",
                                                "example": "01/02/2025 15:30:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Cria um novo assunto",
                "description": "Cria um assunto a partir dos dados fornecidos",
                "operationId": "acfe3773777c814aad62a7c623672b37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string",
                                        "example": "Terror"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Assunto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cod_as": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Comédia"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                }
            }
        },
        "/api/subjects/{cod_as}": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Retorna um assunto pelo Código do Assunto",
                "operationId": "c6c7becc1c49d3d2bfe85d0658f44c99",
                "parameters": [
                    {
                        "name": "cod_as",
                        "in": "path",
                        "description": "Código do assunto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do assunto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cod_as": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Ação"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Atualiza um assunto existente",
                "operationId": "574f503fbf36c9385867cc5aee81ce1c",
                "parameters": [
                    {
                        "name": "cod_as",
                        "in": "path",
                        "description": "Código do assunto a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descricao"
                                ],
                                "properties": {
                                    "descricao": {
                                        "type": "string",
                                        "example": "Ação"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Assunto atualizado com sucesso, sem conteúdo retornado"
                    },
                    "422": {
                        "description": "Falha na validação"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Deleta um assunto pelo Código do Assunto",
                "operationId": "be48e0525e171f57e95f938859a32228",
                "parameters": [
                    {
                        "name": "cod_as",
                        "in": "path",
                        "description": "Código do assunto a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Assunto deletado com sucesso"
                    },
                    "404": {
                        "description": "Assunto não encontrado"
                    }
                }
            }
        }
    },
...
}